fix(server): widen asset_command altitude to 32 bits
The server-side asset_command stored altitude as uint16_t while the wire
message (fss_message_asset_command) and the assets_assetcommand table both
carry 32 bits. An altitude command above 65535 read from the database was
silently truncated before being dispatched to the aircraft.

Widen the asset_command member, constructor parameter and getter to
uint32_t, and widen the ECPG read path (asset_command_s.altitude and the
db_asset_command_get host variable) to unsigned int so the full value
survives from the database to the wire. Add a regression test asserting an
altitude of 100000 round-trips through pack/decode and the session dispatch
path without truncation.

+++ src/client_session.cpp
@@ -49,7 +49,7 @@ auto fss::server::fss_server_details::getPort() -> uint16_t
 
 // NOLINTBEGIN(bugprone-easily-swappable-parameters)
 fss::server::asset_command::asset_command(uint64_t t_dbid, uint64_t t_timestamp, const std::string &t_cmd,
-                                          double t_latitude, double t_longitude, uint16_t t_altitude)
+                                          double t_latitude, double t_longitude, uint32_t t_altitude)
     : dbid(t_dbid), timestamp(t_timestamp), latitude(t_latitude), longitude(t_longitude), altitude(t_altitude)
 // NOLINTEND(bugprone-easily-swappable-parameters)
 {
@@ -111,7 +111,7 @@ auto fss::server::asset_command::getLongitude() -> double
 {
     return this->longitude;
 }
-auto fss::server::asset_command::getAltitude() -> uint16_t
+auto fss::server::asset_command::getAltitude() -> uint32_t
 {
     return this->altitude;
 }
